Continued refactor from moving business logic from TagConfigView to TagProviderRecord
